+++ ncaatourney-svc/src/main/java/com/bray/ncaa/controller/UserController.java
@@ -1,11 +1,12 @@
 package com.bray.ncaa.controller;
 
-import com.bray.ncaa.dao.PoolUserRepository;
 import com.bray.ncaa.model.PoolUser;
 import com.bray.ncaa.model.UserLogin;
+import com.bray.ncaa.service.AdminService;
 import com.bray.ncaa.service.UsersService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
@@ -13,11 +14,20 @@ import org.springframework.web.server.ResponseStatusException;
 import java.util.List;
 
 @RestController
-@RequestMapping("/user")
+@RequestMapping("/api/user")
 @Slf4j
 public class UserController {
     @Autowired
     private UsersService userService;
+
+    @Autowired
+    private AdminService adminService;
+
+    @Value( "${admin.username}" )
+    private String adminUsername;
+
+    @Value( "${admin.password}" )
+    private String adminPassword;
 
     @GetMapping("/list")
     public List<PoolUser> getAllPoolUsers(){
@@ -35,12 +45,14 @@ public class UserController {
     }
 
     @DeleteMapping("/deleteall")
-    public void deleteAllUsers(){
+    public void deleteAllUsers(@RequestHeader String t){
+        adminService.checkAdminAccess(t);
         userService.deleteAllUsers();
     }
 
     @DeleteMapping("/{userID}")
-    public void deleteUser(@PathVariable String userID){
+    public void deleteUser(@PathVariable String userID, @RequestHeader String t){
+        adminService.checkAdminAccess(t);
         userService.deleteUser(userID);
     }
 
@@ -56,6 +68,23 @@ public class UserController {
 
     @PostMapping("login")
     public PoolUser userLogin(@RequestBody UserLogin userLogin) {
+        if(userLogin.getEmail().equals(adminUsername)) {
+            if(userLogin.getPassword().equals(adminPassword)) {
+                PoolUser user = new PoolUser();
+                user.setAdmin(true);
+                user.setAdminToken(adminService.getAdminToken());
+                user.setFirstName("Admin");
+                user.setLastName("User");
+                user.setId("000000000");
+                user.setDisplayName("Admin User");
+
+                return user;
+            } else {
+                throw new ResponseStatusException(
+                        HttpStatus.UNAUTHORIZED, "Admin login incorrect");
+            }
+        }
+
         try {
             return userService.processUserLogin(userLogin);
         } catch (Exception e) {
